requirements.txt
python-telegram-bot
requests
schedule
python-dotenv
openai
discover
numpy
scikit-learn